requirements.txt
asgiref==3.10.0
certifi==2025.11.12
chardet==4.0.0
dj-database-url==3.1.0
Django==5.2.8
django-widget-tweaks==1.5.0
dnspython==2.8.0
gunicorn==23.0.0
idna==2.10
packaging==25.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.6
requests==2.25.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.11.0
